requirements.txt
# Embedding and Vector Store
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Database (Supabase)
supabase>=2.0.0

# API (Phase 2)
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Utilities
pyyaml>=6.0
python-frontmatter>=1.0.0
PyYAML>=6.0
requests>=2.28.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0

# Claude API (for image extraction and thesis system)
anthropic>=0.18.0

# HTTP client (for image fetching)
httpx>=0.25.0
